+++ Assets/Scripts/UnitScales.cs
@@ -4,11 +4,11 @@
 
 public static class UnitScales
 {
-    public static readonly float Mass = 1e10f;
-    public static readonly float Distance = 1e8f;
-    public static readonly float Time = 1e0f;
+    public static readonly double Mass = 1e10;
+    public static readonly double Distance = 1e8;
+    public static readonly double Time = 1e0;
 
-    public static readonly float Velocity = Distance / Time;
-    public static readonly float Force = Mass * (Distance / Time);
-    public static readonly float G = Force * (Distance * Distance) / (Mass * Mass);
+    public static readonly double Velocity = Distance / Time;
+    public static readonly double Force = Mass * (Distance / Time);
+    public static readonly double G = Force * (Distance * Distance) / (Mass * Mass);
 }
